Ignore order of properties unspecified in user array; provide options to change behavior
Now rule-properties-order will by default ignore the order of unspecified properties
if the user has provided an array. This is a change to the default behavior previously,
so might be considered a breaking change.

The `unspecified` option has been added for the user to explicitly decide where
unspecified properties should go: "top", "bottom", or "ignore" (same as
default behavior).

Closes #374.