Add caveats section
Alright, I'm seriously not trying to be a dick with this one, and this is literally the last time I come to the stylelint repo for weeks/months.

I think this PR is a nice compromise that perhaps all projects should do. Just provide a little caveats section toward the bottom of the README so people know exactly what they're getting themselves into.

If you don't want to merge because you don't like me, I get that and I'm sorry for offending everyone, but please look past that for future users who haven't done anything.