Enable Code Climate duplication engine
- Ignore similar code issues
- Tune mass threshold
- Exclude `system-tests/*`+`**/__tests__/*`

The Code Climate duplication engine today may emit issues for similar
argument/paramater/import structures in JavaScript, which makes this
engine less useful. This is particurly unhelpful when the project has a
lot of intentional similar structure, such as a static analysis tool
with a common rule structure.

Until Code Climate implements better handling for these structures in
JavaScript, the engine is best configured with a higher mass threshold,
limited to identical duplication issues, and not run on test files,
which also often have higher occurrences of intentional duplicate
structures.