Alias *-whitelist, *-blacklist and *-requirelist rules
Previously *-whitelist, *-blacklist and *-requirelist rules were renamed
to a consistent naming pattern of *-allowed-list, *-disallowed-list and
*-required-list respectively. This commit adds aliases so that the old
rules can continue to operate; this allows backwards compatibility to be
maintained.

A consequence of this is that violations for the previously named rules
will show up with the new name. A developer may also get some confusing
results if they happen to have the same rule defined with both names and
different options.