Remove `postcss-reporter` package
This change removes the `postcss-reporter` package and reduces the dependencies.
As you see the changed `package-lock.json`, this package depends on:

- `chalk`
- `lodash`
- `log-symbols`
- `postcss`

See also https://github.com/stylelint/stylelint/issues/2454#issuecomment-622604821

We've used only the tiny utility function `getLocation()` in the package.
This function just returns an object with the `line` and `column` property in our project,
so I think we can replace it with a simple code.

https://github.com/postcss/postcss-reporter/blob/f01a601ea2cd41d626e561969d66a765b3afcb2d/lib/util.js#L4-L13