Add type definition for `postcss-syntax`
This change adds a type definition for the `postcss-syntax` package.
To find the definition easily, this extracts a new file from the `types/postcss/index.d.ts` file.

Note that there is no `@types/postcss-syntax` package.
(see <https://www.npmjs.com/search?q=%40types%2Fpostcss-syntax>)

Here is a main source of the package:
https://github.com/gucong3000/postcss-syntax/blob/v0.36.2/index.js

Also, `postcss-syntax` is used only in `lib/getPostcssResult.js`:

https://github.com/stylelint/stylelint/blob/cecacb9d4fe3b5bae5f264efa6e60c414a0f4946/lib/getPostcssResult.js#L64-L75

This is a part of #4399.