Re-instate *-blacklist/*-requirelist/*-whitelist rules
The aliasing approach taken for these rule renaming presented user
experience problems [1]. To resolve these problems the preference is to
revert on the aliasing strategy and instead return to a deprecation
strategy. The first step towards this is re-instating these rules.

As per the preference of jeddy3 [2] these rules are re-instated using a copy
and paste strategy, with the expectation these will be removed in the
next major release. An approach that didn't involve copy and pasting was
previously introduced in [3] and could be looked at again if maintaining this
duplicate code proves problematic.

[1]: https://github.com/stylelint/stylelint/issues/4854#issuecomment-654332844
[2]: https://github.com/stylelint/stylelint/issues/4854#issuecomment-670935654
[3]: https://github.com/stylelint/stylelint/commit/e93e44c938d77021eec4893fc20327337a60d49f