Deprecation links point to GitHub tag
This reflects the documentation on deprecating stylelint rules [1] by
linking to the GitHub website rather than stylelint website, so that the
link can continue operating indefinitely.

As suggested in the PR [2] these have been linked to the anticipated next
release of stylelint, so these links will not be actually operational
until this tag is made.

[1]: https://github.com/stylelint/stylelint/blob/858dcd584224042654d80ce8fa8ad71f41f20808/docs/developer-guide/rules.md#deprecate-a-rule
[2]: https://github.com/stylelint/stylelint/pull/4892#issuecomment-679375149