Deprecate *-blacklist/*-requirelist/*-whitelist (#4892)
* Make copies for *-blacklist/*-requirelist/*-whitelist rules

As discussed in issue #4854, the desired approach is to deprecate the
*-blacklist/*-requirelist/*-whitelist rules. The first step I've taken
to doing this is to make copies of all the rules. The commands used to
generate this commit were:

- find . -name '*-allowed-list' -exec sh -c 'cp -R "$1" "${1%-allowed-list}-whitelist"' _ {} \;
- find . -name '*-disallowed-list' -exec sh -c 'cp -R "$1" "${1%-disallowed-list}-blacklist"' _ {} \;
- find . -name '*-required-list' -exec sh -c 'cp -R "$1" "${1%-required-list}-requirelist"' _ {} \;

* Re-instate *-blacklist/*-requirelist/*-whitelist rules

The aliasing approach taken for these rule renaming presented user
experience problems [1]. To resolve these problems the preference is to
revert on the aliasing strategy and instead return to a deprecation
strategy. The first step towards this is re-instating these rules.

As per the preference of jeddy3 [2] these rules are re-instated using a copy
and paste strategy, with the expectation these will be removed in the
next major release. An approach that didn't involve copy and pasting was
previously introduced in [3] and could be looked at again if maintaining this
duplicate code proves problematic.

[1]: https://github.com/stylelint/stylelint/issues/4854#issuecomment-654332844
[2]: https://github.com/stylelint/stylelint/issues/4854#issuecomment-670935654
[3]: https://github.com/stylelint/stylelint/commit/e93e44c938d77021eec4893fc20327337a60d49f

* Deprecate *-blacklist/*-requirelist/*-whitelist rules

This applies deprecation warnings when these rules are used, tests to
check there is a deprecation warning and re-includes them in
documentation.

* Deprecation links point to GitHub tag

This reflects the documentation on deprecating stylelint rules [1] by
linking to the GitHub website rather than stylelint website, so that the
link can continue operating indefinitely.

As suggested in the PR [2] these have been linked to the anticipated next
release of stylelint, so these links will not be actually operational
until this tag is made.

[1]: https://github.com/stylelint/stylelint/blob/858dcd584224042654d80ce8fa8ad71f41f20808/docs/developer-guide/rules.md#deprecate-a-rule
[2]: https://github.com/stylelint/stylelint/pull/4892#issuecomment-679375149