Removed "hasMagic" check before escaping paths.
Globby's "hasMagic" function, which is meant to detect if a path
contains globbing characters, notably does not detect paths that
contain matched parentheses, unless those matched parentheses
have extra globbing characters attached to them. This is a problem
because globby requires matches parentheses to be escaped, if they
are part of the literal path.
This changes the part of the code that escapes paths that literally
exist to not check "hasMagic" first, to get around that oversight.