Add configuration example at end of docs
A proposal to add configuration examples to the docs for the various rules. It is not often that I need to update or change the default shareable config, but when I do, oh boy 😄 I can never remember how the various rules are configured. As such, I think it would be useful to add little examples like this one to rules that are not a simple `true|false`.

Also, I believe in this case you set the rule to `null` to disable?