Validate color-hex-case with postcss-value-parser
As discussed in #4826, `style-search` has false positives when used with
`css-in-js` object notation. This change updates the `color-hex-case`
rule to use `postcss-value-parser` instead, which has support for
`css-in-js` as a syntax.

The unit tests now also include a suite for testing both `upper` and
`lower` options for `color-hex-case` with the `css-in-js` syntax.