draft: fix end positions for `no-duplicate-selectors`
Unsure of what the highlighted positions should be / what's the most
useful for the VSCode extension. Right now, it just highlights the
first duplicate selector. I believe this is because of the
implementation of the `word` parsing. Is it more useful to highlight
each duplicate selector? The last one? etc.