Fix end positions for `no-duplicate-*` (#6047)
* draft: fix end positions for `no-duplicate-selectors`

Unsure of what the highlighted positions should be / what's the most
useful for the VSCode extension. Right now, it just highlights the
first duplicate selector. I believe this is because of the
implementation of the `word` parsing. Is it more useful to highlight
each duplicate selector? The last one? etc.

* draft: fix end positions for `no-duplicate-at-import-rules`

Two concerns:
- not sure if calling `atRule.toString()` is the appropriate use;
  should I be using `raws`?
- in this case, my suggested start/end positions are larger than the
  message rejection, but it aims to match the positions indicated
  in the README