Fix Unexpected Node API output when any rule contains `disableFix: true` #6542
fixes #6542 and bv that should also fix https://github.com/stylelint/vscode-stylelint/issues/369

Issue: The API did not return the fixed css in `output` if any rule contained `ruleDisableFix`.

I have removed the `ruleDisableFix` code, because it did not affect the code fixing itself and was assumingly wrongly used to change the content format of `output`.
I have adjusted some tests to adapt to the changed behaviour and extended the test conditions to check that the returned code has the correct fixes applied.

Issue was introduced in https://github.com/stylelint/stylelint/pull/5460