Add `allowEmptyInput`, `cache`, `fix` options to configuration object (#6778)
Part of (but does not close) #5142. This PR handles all the `boolean` flags (which should be relatively simple).

> Is there anything in the PR that needs further explanation?

Sorry about the CodeQL alerts - it was an approach that I tried (and chose to abandon) that I forgot to remove 😅 

As well as my note below on `cache`, two quick questions:

1. For the configuration docs, what orders should the options go in? I may have missed a pattern - I was confused since it's not in the same order as the flags.
2. Any tests I should add?

---

Note that I didn't `augmentConfig` for `allowEmptyInput` and `cache`, since we never access them through the `config` object outside of `standalone.js`; including them drops code coverage. That being said, maybe there's a cleaner way to use these options "downstream", and be consistent in augmenting the config?
---------

Co-authored-by: Masafumi Koba <473530+ybiquitous@users.noreply.github.com>