Fix `declaration-block-no-redundant-longhand-properties` false negative for `font-synthesis` (#7214)
Note: `font-synthesis` actually supports *four* longhand properties. However, `font-synthesis-position` currently has [**0.41%** adoption according to caniuse](https://caniuse.com/mdn-css_properties_font-synthesis-position), which is far too small for it to be relevant. I've elected to not include it for now; however, we could also include it if we think that better future-proofs the rule? 

---------

Co-authored-by: Masafumi Koba <473530+ybiquitous@users.noreply.github.com>