Fix type definitions for CommonJS (#7377)
This fixes the type definitions for CJS by leveraging a namespace. ESM
type definitions were added, which export all types from the CJS types,
and exports the exported stylelint object as default.

Some tweaks were made to the package.json exports, to resolve the types
correctly for both CJS and ESM.

The type test was modified to fix CJS usage. Also a test was added for
ESM usage. Stylelint was using tsconfig `paths`, which is a commonly
misunderstood / misused feature that should rarely be used. This was
removed, so TypeScript resolves the types based on package exports.

Co-authored-by: Masafumi Koba <473530+ybiquitous@users.noreply.github.com>