Fix report flags not reporting on subsequent runs when cache is used
The bug's reason is that the `stylelintError`/`stylelintWarning` properties
in a lint result object are not set when `styelint-disable` comment problems are reported.

Also, this change refactors the code around `styelint-disable` comments
to make it more readable and prevent possible bugs.

The new code uses [`Result#warn()`](https://postcss.org/api/#result-warn) in PostCSS for a comment node,
like reporting a normal rule problem.