Fix `benchmark-rule` script to resolve `TypeError`
This fixes the following `TypeError` by using the Stylelint Public API,
instead of PostCSS and Stylelint internals.

```
TypeError: Cannot read properties of undefined (reading 'config')
    at validateOptions (file:///Users/masafumi/git/stylelint/stylelint/lib/utils/validateOptions.mjs:19:40)
```

This makes the benchmark script slower, while the script will be less fragile.

In addition, this converts the script's third argument from a rule context to a lint config
because the `context.fix` property is now deprecated.